0.20.1·hotfix：修复书封取色符号位 bug导致开屏闪退
MetaCoverPalette.of 在 FNV 哈希高位为 1 时先强转 Int 得到负数，
再 % PALETTES.size 产生负索引，任何带封面的界面冷启动即
ArrayIndexOutOfBounds 闪退。改为先对 Long 取模再转 Int（恒非负）。
versionCode 107->109，versionName->0.20.1；APK 14MB 延续 0.20.0 精简。

+++ app/src/main/java/com/xiaoswz/reader/ui/theme/MetaCoverPalette.kt
@@ -56,7 +56,8 @@ object MetaCoverPalette {
      */
     fun of(seed: String): CoverPalette {
         if (seed.isBlank()) return PALETTES[0]
-        val idx = ((hash(seed) and 0xFFFFFFFFL).toInt()) % PALETTES.size
+        // 先对 Long 取模（结果恒为非负），再转 Int，避免高位被当作负数导致负索引
+        val idx = ((hash(seed) and 0xFFFFFFFFL) % PALETTES.size.toLong()).toInt()
         return PALETTES[idx]
     }
